add implementation for Book repository. Fix some bugs

+++ BookStorage.Application/Implementations/Services/AuthorService.cs
@@ -26,7 +26,7 @@ namespace BookStorage.Application.Implementations.Services
 		public async Task<Author?> Create(Author author)
 		{
 			var isExist = await _repository.CheckIfAuthorAlreadyExistsAsync(author.FirstName, author.LastName, author.BirthDate);
-			if (!isExist)
+			if (isExist)
 				return null;
 
 			return await _repository.CreateAsync(author);
@@ -34,20 +34,18 @@ namespace BookStorage.Application.Implementations.Services
 
 		public async Task<Author?> Update(Guid id, Author author)
 		{
-			//var existingAuthor = Authors.FirstOrDefault(a => a.Id == id);
-			//if (existingAuthor == null)
-			//	return null;
+			var existingAuthor = await _repository.GetByIdAsync(id);
+			if (existingAuthor == null)
+				return null;
 
 			return await _repository.UpdateAsync(id, author);
 		}
 
 		public async Task<bool> Delete(Guid id)
 		{
-			//var authorToDelete = Authors.FirstOrDefault(a => a.Id == id);
-			//if (authorToDelete == null)
-			//	return false;
-
-			//bool wasDeletedAuthor = Authors.Remove(authorToDelete);
+			var existingAuthor = await _repository.GetByIdAsync(id);
+			if (existingAuthor == null)
+				return false;
 
 			return await _repository.DeleteAsync(id);
 		}
